requirements.txt
django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.0 